build: Switch from boot to leiningen
It's a shame. Boot is really cool, but for a small project like this it
seems to be more pain than it's worth. In particular I am finding
clojars deployment to be a bit of a struggle.

Back to the magic of leiningen :( Not that leiningen is awful, but I
really do think boot has the right idea: defining build processes,
rather than defining projects.